refactor: trocear GruposScreen en components/grupos (parcial, Fase 1.7)
1100 → 561 líneas (sale de la lista de >800). Extraído a components/grupos/:
SearchBar, MemberRow, GrupoCard, SearchHitRow (ya recibían styles por prop),
gruposStyles.ts (createStyles + tipo GruposStyles) y gruposHelpers.ts (tipos +
normalize/isMe, con test). highlightText movido a SearchHitRow (único uso).

Cero cambios de comportamiento. Parcial a propósito: las 4 ramas de render del
cuerpo (~511 líneas) no se tocan (requiere verificación en dispositivo; la
migración a SectionList cambiaría comportamiento). Gigantes >800: 11 → 10.
typecheck/typecheck:tests/lint(0 err)/test(204) ok.
