+++ docs/planes/PLAN_CALIDAD.md
@@ -81,8 +81,8 @@ Objetivo: ninguna pantalla > 800 líneas; cada pantalla = composición de compon
 - [ ] **1.4 — `contigo/evangelio.tsx` (1.295) + `contigo/oracion.tsx` (1.170)** — mismo dominio; sospecha fuerte de duplicación entre ambas. **Antes de trocear, comparar**: lo común va a `components/contigo/` y a un hook compartido; después trocear lo que quede.
 - [ ] **1.5 — `calendario.tsx` (1.123)**.
 - [ ] **1.6 — `SecretPanelModal.tsx` (1.121)** — panel admin; trocear por sección del panel.
-- [ ] **1.7 — `GruposScreen.tsx` (1.100)** — tiene 4 variantes de render con `.map()` anidados (ver MEJORAS §1.4); al trocear, migrar a `SectionList` con `GrupoItem`/`MiembroRow` memoizados (mata dos pájaros).
-- [ ] **1.8 — `HomeWidgets.tsx` (1.075)** — un archivo con N widgets dentro; partirlo en un archivo por widget.
+- [~] **1.7 — `GruposScreen.tsx` (1.100)** — **parcial, 2026-06-28** → 561 líneas (fuera de la lista de gigantes). Extraídos a `components/grupos/`: los subcomponentes (`SearchBar`, `MemberRow`, `GrupoCard`, `SearchHitRow`), los estilos (`gruposStyles.ts`) y los helpers puros (`gruposHelpers.ts`: `normalize`/`isMe`/tipos, con test). **Pendiente** (no tocado a propósito, requiere verificación en dispositivo): trocear las 4 ramas de render del cuerpo principal (~511 líneas, sigue >400) y valorar la migración a `SectionList` con items memoizados (MEJORAS §1.4) — eso sí cambia comportamiento.
+- [ ] **1.8 — `HomeWidgets.tsx` (1.075)** — un archivo con N widgets dentro; partirlo en un archivo por widget. ⚠️ Ojo: hay una extracción previa **huérfana** (`components/ui/ProgressRing.tsx`, `StatCard.tsx`, `TeaserCard.tsx` "extracted from HomeWidgets" pero **no las importa nadie** — Contigo sigue usando las de HomeWidgets). Antes de trocear hay que decidir si se adoptan esas versiones de `ui/` o se borran (dedup = cambio de comportamiento), si no se duplica más el lío.
 - [ ] **1.9 — Segunda ronda** (si hay energía): `notifications.tsx` (994), ~~`EvaluationWizard.tsx` (974)~~ **hecho 2026-06-28** → 360 líneas; extraídas a `components/evaluation/` las fases (`WelcomePhase`, `QuestionInput`, `SuccessPhase`), los controles (`WizardButton`, `ScaleInput`) y los estilos del armazón (`wizardStyles.ts`). ~~`NotificationsBottomSheet.tsx` (937)~~ **hecho 2026-06-28** → 365 líneas; extraídos a `components/notifications/` el detalle (`NotificationDetail`), el item de lista (`NotificationListItem`) y los helpers puros (`notificationDisplay.ts`, con test). ~~`PreviewChannelModal.tsx` (847)~~ **hecho 2026-06-28** → 349 líneas; piezas decorativas extraídas a `components/preview-channel/` (`AnimatedGradients`, `FloatingParticle`, `ConfettiBurst`, `GiantLever`, `LabDecorations`).
   - Nota: `hooks/useSongProcessor.ts` pasó a >800 (1014) con el trabajo del cantoral (pantalla de error de sintaxis), que compensa parte de estos troceados en el recuento de gigantes (hoy 11).
